pyproject.toml
[tool.poetry]
name = "cryptobotkit"
version = "0.1.0"
description = ""
authors = ["Owen Campbell <owen.campbell@tanti.org.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^4.3"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
